requirements.txt
agate==1.9.1
aiosqlite==0.22.1
alembic==1.18.4
altair==6.0.0
amplitude-analytics==1.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apify_client==2.5.0
apify_shared==2.2.0
apprise==1.9.8
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asyncpg==0.31.0
attrs==25.4.0
babel==2.18.0
beartype==0.22.9
blinker==1.9.0
boto3==1.42.69
botocore==1.42.69
cachetools==7.0.5
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==4.1.0
cronsim==2.7
cryptography==46.0.5
cyclopts==4.10.0
daff==1.4.2
dateparser==1.3.0
dbt-adapters==1.22.9
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-protos==1.0.440
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.11.3
deepdiff==8.6.1
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
exceptiongroup==1.3.1
fakeredis==2.34.1
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
impit==0.12.0
importlib_metadata==8.7.1
iniconfig==2.3.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.1
jaraco.functools==4.4.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
leather==0.4.1
lupa==2.6
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
oauthlib==3.3.1
opentelemetry-api==1.40.0
orderly-set==5.5.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
prefect==3.6.22
prometheus_client==0.24.1
protobuf==6.33.5
py-key-value-aio==0.4.4
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.18.2
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==25.3.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
readchar==4.2.1
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
s3transfer==0.16.0
semver==3.0.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
snowflake-connector-python==4.3.0
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sqlparse==0.5.4
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
text-unidecode==1.3
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
whenever==0.9.5
zipp==3.23.0
